Trabalhando em cima do cadastro de produtos e cotações

+++ Tp3DR3/src/main/java/br/edu/infnet/Tp3DR3/model/repository/ICotacaoRepository.java
@@ -7,6 +7,6 @@ import br.edu.infnet.Tp3DR3.model.negocio.Cotacao;
 
 
 @Repository
-public interface ICotacaoRepository extends JpaRepository<Cotacao, Short> {
+public interface ICotacaoRepository extends JpaRepository<Cotacao, Integer> {
 	
 }
