+++ Tp3DR3/src/main/java/br/edu/infnet/Tp3DR3/model/repository/IProdutoRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import br.edu.infnet.Tp3DR3.model.negocio.Produto;
 
 
-public interface IProdutoRepository extends JpaRepository<Produto, Short> {
+public interface IProdutoRepository extends JpaRepository<Produto, Integer> {
 	@Query(value = "SELECT p FROM Produto p WHERE p.nomeProduto LIKE '%' || :keyword || '%'")
     public Produto search(@Param("keyword") String keyword);
 	public List<Produto> findCotacoes();
